+++ src/onegov.gazette/onegov/gazette/tests/test_views_dashboard.py
@@ -42,6 +42,9 @@ def test_view_dashboard(gazette_app):
         manage.form['category'] = '11'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
         manage.form.submit()
 
         manage = editor_1.get('/').maybe_follow()
@@ -182,6 +185,9 @@ def test_view_dashboard(gazette_app):
         manage.form['category'] = '11'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
         manage.form.submit()
 
         manage = editor_1.get('/').maybe_follow()  # other group