+++ src/onegov.gazette/onegov/gazette/tests/test_views_notice.py
@@ -24,6 +24,9 @@ def test_view_notice(gazette_app):
             manage.form['billing_address'] = 'someone\nstreet\r\nplace'
             manage.form['issues'] = ['2017-44', '2017-45']
             manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
             manage.form.submit()
 
         # check if the notices are displayed correctly
@@ -36,6 +39,8 @@ def test_view_notice(gazette_app):
                 view = user.get('/notice/{}'.format(slug))
                 assert title in view
                 assert "1. Oktober 2017" in view
+                assert "Govikon, 1. Januar 2019" in view
+                assert "State Chancellerist" in view
                 assert "Civic Community" in view
                 assert "Education" in view
                 assert "<dd>Ja</dd>" in view
@@ -83,6 +88,9 @@ def test_view_notice_actions(gazette_app):
             manage.form['category'] = '11'
             manage.form['issues'] = ['2017-44']
             manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
             manage.form.submit()
 
         # check the actions
@@ -255,11 +263,16 @@ def test_view_notice_preview(gazette_app):
         manage.form['category'] = '11'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
         manage.form.submit()
 
     view = editor_1.get('/notice/titel/preview')
     assert "Titel" in view
     assert "1. Oktober 2017" in view
+    assert "Govikon, 1. Januar 2019" in view
+    assert "State Chancellerist" in view
     assert "Civic Community" not in view
     assert "Education" not in view
     assert "TestGroup" not in view
@@ -267,6 +280,214 @@ def test_view_notice_preview(gazette_app):
     assert "Nr. 45, 10.11.2017" not in view
     assert "in Arbeit" not in view
     assert "erstellt" not in view
+
+
+def test_view_notice_submit(gazette_app):
+    editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
+
+    with freeze_time("2017-11-01 11:00"):
+        # create a notice for each editor
+        for count, user in enumerate((editor_1, editor_2, editor_3)):
+            manage = user.get('/notices/drafted/new-notice')
+            manage.form['title'] = 'Titel {}'.format(count + 1)
+            manage.form['organization'] = '200'
+            manage.form['category'] = '11'
+            manage.form['issues'] = ['2017-44', '2017-45']
+            manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
+            manage.form.submit()
+
+        # check invalid actions
+        for action in ('reject', 'accept'):
+            assert not publisher.get('/notice/titel-1/{}'.format(action)).forms
+            assert not publisher.get('/notice/titel-2/{}'.format(action)).forms
+            assert not publisher.get('/notice/titel-3/{}'.format(action)).forms
+            for user in (editor_1, editor_2, editor_3):
+                user.get('/notice/titel-1/{}'.format(action), status=403)
+                user.get('/notice/titel-2/{}'.format(action), status=403)
+                user.get('/notice/titel-3/{}'.format(action), status=403)
+
+        # check if invalid permissions
+        submit_notice(editor_1, 'titel-3', forbidden=True)
+        submit_notice(editor_2, 'titel-3', forbidden=True)
+        submit_notice(editor_3, 'titel-1', forbidden=True)
+        submit_notice(editor_3, 'titel-2', forbidden=True)
+
+    # check deadlines (2 gets submitted)
+    with freeze_time("2017-11-08 13:00"):
+        submit_notice(editor_1, 'titel-2', redirected=True)
+        submit_notice(publisher, 'titel-2', redirected=True)
+    with freeze_time("2017-11-01 13:00"):
+        submit_notice(editor_1, 'titel-2', redirected=True)
+        submit_notice(publisher, 'titel-2')
+
+    # check invalid category (1 gets submitted)
+    with freeze_time("2017-11-01 11:00"):
+        change_category(gazette_app, '11', active=False)
+        submit_notice(editor_1, 'titel-1', redirected=True)
+        submit_notice(publisher, 'titel-1', redirected=True)
+
+        change_category(gazette_app, '11', active=True)
+        submit_notice(editor_1, 'titel-1')
+
+    # check invalid organization (3 gets submitted)
+    with freeze_time("2017-11-01 11:00"):
+        change_organization(gazette_app, '200', active=False)
+        submit_notice(editor_3, 'titel-3', redirected=True)
+        submit_notice(publisher, 'titel-3', redirected=True)
+
+        change_organization(gazette_app, '200', active=True)
+        submit_notice(editor_3, 'titel-3')
+
+
+def test_view_notice_reject(gazette_app):
+    editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
+
+    with freeze_time("2017-11-01 11:00"):
+        # create a notice for each editor
+        for count, user in enumerate((editor_1, editor_2, editor_3)):
+            manage = user.get('/notices/drafted/new-notice')
+            manage.form['title'] = 'Titel {}'.format(count + 1)
+            manage.form['organization'] = '200'
+            manage.form['category'] = '11'
+            manage.form['issues'] = ['2017-44', '2017-45']
+            manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
+            manage.form.submit()
+            submit_notice(user, 'titel-{}'.format(count + 1))
+
+        # check wrong actions
+        submit_notice(publisher, 'titel-1', unable=True)
+        submit_notice(publisher, 'titel-2', unable=True)
+        submit_notice(publisher, 'titel-3', unable=True)
+        submit_notice(editor_1, 'titel-1', unable=True)
+        submit_notice(editor_1, 'titel-2', unable=True)
+        submit_notice(editor_1, 'titel-3', forbidden=True)
+        submit_notice(editor_2, 'titel-1', unable=True)
+        submit_notice(editor_2, 'titel-2', unable=True)
+        submit_notice(editor_2, 'titel-3', forbidden=True)
+        submit_notice(editor_3, 'titel-1', forbidden=True)
+        submit_notice(editor_3, 'titel-2', forbidden=True)
+        submit_notice(editor_3, 'titel-3', unable=True)
+
+        # check if the notices can be rejected
+        for user, slug, forbidden in (
+            (editor_1, 'titel-1', True),
+            (editor_1, 'titel-2', True),
+            (editor_1, 'titel-3', True),
+            (editor_3, 'titel-3', True),
+            (publisher, 'titel-1', False),
+            (publisher, 'titel-2', False),
+            (publisher, 'titel-3', False),
+        ):
+            reject_notice(user, slug, forbidden=forbidden)
+
+
+def test_view_notice_accept(gazette_app):
+    editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
+
+    with freeze_time("2017-11-01 11:00"):
+        # create a notice for each editor
+        for count, user in enumerate((editor_1, editor_2, editor_3)):
+            manage = user.get('/notices/drafted/new-notice')
+            manage.form['title'] = 'Titel {}'.format(count + 1)
+            manage.form['organization'] = '410'
+            manage.form['category'] = '11'
+            manage.form['at_cost'].select('yes')
+            manage.form['billing_address'] = 'someone\nstreet\nplace'
+            manage.form['issues'] = ['2017-44', '2017-45']
+            manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
+            manage.form.submit()
+            submit_notice(user, 'titel-{}'.format(count + 1))
+
+        # check wrong actions
+        submit_notice(publisher, 'titel-1', unable=True)
+        submit_notice(publisher, 'titel-2', unable=True)
+        submit_notice(publisher, 'titel-3', unable=True)
+        submit_notice(editor_1, 'titel-1', unable=True)
+        submit_notice(editor_1, 'titel-2', unable=True)
+        submit_notice(editor_1, 'titel-3', forbidden=True)
+        submit_notice(editor_2, 'titel-1', unable=True)
+        submit_notice(editor_2, 'titel-2', unable=True)
+        submit_notice(editor_2, 'titel-3', forbidden=True)
+        submit_notice(editor_3, 'titel-1', forbidden=True)
+        submit_notice(editor_3, 'titel-2', forbidden=True)
+        submit_notice(editor_3, 'titel-3', unable=True)
+
+    # check redirect for past issues
+    with freeze_time("2017-11-04 11:00"):
+        accept_notice(publisher, 'titel-1', redirected=True)
+
+    # check redirect for invalid category
+    with freeze_time("2017-11-04 11:00"):
+        change_category(gazette_app, '11', active=False)
+        accept_notice(publisher, 'titel-1', redirected=True)
+        change_category(gazette_app, '11', active=True)
+
+    # check redirect for invalid organization
+    with freeze_time("2017-11-01 11:00"):
+        change_organization(gazette_app, '410', active=False)
+        accept_notice(publisher, 'titel-1', redirected=True)
+        change_organization(gazette_app, '410', active=True)
+
+    with freeze_time("2017-11-01 15:00"):
+        # check if the notices can be accepted
+        for user, slug, forbidden in (
+            (editor_1, 'titel-1', True),
+            (editor_1, 'titel-2', True),
+            (editor_1, 'titel-3', True),
+            (editor_3, 'titel-3', True),
+            (publisher, 'titel-1', False),
+            (publisher, 'titel-2', False),
+            # (publisher, 'titel-3', False),
+        ):
+            accept_notice(user, slug, forbidden=forbidden)
+
+        message = gazette_app.smtp.outbox.pop()
+        assert message['From'] == 'mails@govikon.ch'
+        assert message['To'] == 'printer@onegov.org'
+        assert message['Reply-To'] == 'mails@govikon.ch'
+        assert '44  Titel 2' in message['Subject']
+        payload = message.get_payload(1).get_payload(decode=True)
+        payload = payload.decode('utf-8')
+        assert '44  Titel 2' in payload
+        assert "Govikon, 1. Januar 2019" in payload
+        assert "State Chancellerist" in payload
+        assert "Kostenpflichtig 85% vom Normalpreis" in payload
+        assert "someone<br>street<br>place" in payload
+
+        message = gazette_app.smtp.outbox.pop()
+        assert message['From'] == 'mails@govikon.ch'
+        assert message['To'] == 'printer@onegov.org'
+        assert message['Reply-To'] == 'mails@govikon.ch'
+        assert '44  Titel 1' in message['Subject']
+        payload = message.get_payload(1).get_payload(decode=True)
+        payload = payload.decode('utf-8')
+        assert '44  Titel 1' in payload
+
+        principal = gazette_app.principal
+        principal.publish_from = 'publisher@govikon.ch'
+        gazette_app.cache.set('principal', principal)
+
+        change_organization(gazette_app, '400', external_name='xxx')
+
+        accept_notice(publisher, 'titel-3')
+
+        message = gazette_app.smtp.outbox.pop()
+        assert message['From'] == 'mails@govikon.ch'
+        assert message['To'] == 'printer@onegov.org'
+        assert message['Reply-To'] == 'publisher@govikon.ch'
+        assert '44 xxx Titel 3' in message['Subject']
+        payload = message.get_payload(1).get_payload(decode=True)
+        payload = payload.decode('utf-8')
+        assert '44 xxx Titel 3' in payload
 
 
 def test_view_notice_delete(gazette_app):
@@ -281,6 +502,9 @@ def test_view_notice_delete(gazette_app):
             manage.form['category'] = '11'
             manage.form['issues'] = ['2017-44', '2017-45']
             manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
             manage.form.submit()
 
             manage = user.get('/notice/erneuerungswahlen/delete')
@@ -295,6 +519,9 @@ def test_view_notice_delete(gazette_app):
             manage.form['category'] = '11'
             manage.form['issues'] = ['2017-44', '2017-45']
             manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
             manage.form.submit()
 
             submit_notice(user, 'erneuerungswahlen')
@@ -313,6 +540,9 @@ def test_view_notice_delete(gazette_app):
             manage.form['category'] = '11'
             manage.form['issues'] = ['2017-44', '2017-45']
             manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
             manage.form.submit()
 
             submit_notice(user, 'erneuerungswahlen')
@@ -330,6 +560,9 @@ def test_view_notice_delete(gazette_app):
             manage.form['category'] = '11'
             manage.form['issues'] = ['2017-44', '2017-45']
             manage.form['text'] = "1. Oktober 2017"
+            manage.form['author_place'] = 'Govikon'
+            manage.form['author_name'] = 'State Chancellerist'
+            manage.form['author_date'] = '2019-01-01'
             manage.form.submit()
 
             submit_notice(editor_1, 'erneuerungswahlen')
@@ -370,6 +603,9 @@ def test_view_notice_changelog(gazette_app):
         manage.form['category'] = '11'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
         manage.form.submit()
 
     with freeze_time("2017-11-01 11:02"):
@@ -420,6 +656,233 @@ def test_view_notice_changelog(gazette_app):
     ]
 
 
+def test_view_notice_edit(gazette_app):
+    editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
+
+    with freeze_time("2017-11-01 11:00"):
+        manage = editor_1.get('/notices/drafted/new-notice')
+        manage.form['title'] = "Notice"
+        manage.form['organization'] = '200'
+        manage.form['category'] = '11'
+        manage.form['issues'] = ['2017-44', '2017-45']
+        manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
+        manage.form.submit()
+
+        edit_notice(
+            editor_1, 'notice',
+            title="Kantonsratswahl",
+            organization='300',
+            category='12',
+            issues=['2017-46'],
+            text="1. Dezember 2017",
+        )
+        view = editor_1.get('/notice/notice')
+        assert "Kantonsratswahl" in view
+        assert "Municipality" in view
+        assert "Submissions" in view
+        assert "Nr. 46, 17.11.2017" in view
+        assert "1. Dezember 2017" in view
+
+        # drafted
+        for user, title, forbidden in (
+            (editor_1, 'XXX1', False),
+            (editor_2, 'XXX2', False),
+            (editor_3, 'XXX3', True),
+            (publisher, 'XXX4', False),
+        ):
+            edit_notice(user, 'notice', title=title, forbidden=forbidden)
+            if not forbidden:
+                assert title in editor_1.get('/notice/notice')
+
+        # submitted
+        submit_notice(editor_1, 'notice')
+        for user, title, forbidden in (
+            (editor_1, 'YYY1', False),
+            (editor_2, 'YYY2', False),
+            (editor_3, 'YYY3', True),
+            (publisher, 'YYY4', False),
+        ):
+            edit_notice(user, 'notice', title=title, forbidden=forbidden)
+            if not forbidden:
+                assert title in editor_1.get('/notice/notice')
+
+        # rejected
+        reject_notice(publisher, 'notice')
+        for user, title, forbidden in (
+            (editor_1, 'YYY1', False),
+            (editor_2, 'YYY2', False),
+            (editor_3, 'YYY3', True),
+            (publisher, 'YYY4', False),
+        ):
+            edit_notice(user, 'notice', title=title, forbidden=forbidden)
+            if not forbidden:
+                assert title in editor_1.get('/notice/notice')
+
+        # accepted
+        submit_notice(editor_1, 'notice')
+        accept_notice(publisher, 'notice')
+        edit_notice(editor_1, 'notice', unable=True)
+        edit_notice(editor_2, 'notice', unable=True)
+        edit_notice(editor_3, 'notice', forbidden=True)
+        edit_notice(publisher, 'notice', unable=True)
+
+
+def test_view_notice_edit_deadlines(gazette_app):
+    editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
+
+    with freeze_time("2017-11-01 11:00"):
+        manage = editor_1.get('/notices/drafted/new-notice')
+        manage.form['title'] = "Notice"
+        manage.form['organization'] = '200'
+        manage.form['category'] = '11'
+        manage.form['issues'] = ['2017-44', '2017-45']
+        manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
+        manage.form.submit()
+
+    marker = "Bitte Ausgaben neu wählen."
+    with freeze_time("2017-11-01 11:00"):
+        assert marker not in editor_1.get('/notice/notice/edit')
+        assert marker not in publisher.get('/notice/notice/edit')
+
+    with freeze_time("2017-11-01 13:00"):
+        assert marker in editor_1.get('/notice/notice/edit')
+        assert marker not in publisher.get('/notice/notice/edit')
+
+    with freeze_time("2017-11-10 13:00"):
+        assert marker in editor_1.get('/notice/notice/edit')
+        assert marker in publisher.get('/notice/notice/edit')
+
+
+def test_view_notice_edit_invalid_category(gazette_app):
+    editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
+
+    with freeze_time("2017-11-01 11:00"):
+        manage = editor_1.get('/notices/drafted/new-notice')
+        manage.form['title'] = "Notice"
+        manage.form['organization'] = '200'
+        manage.form['category'] = '11'
+        manage.form['issues'] = ['2017-44', '2017-45']
+        manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
+        manage.form.submit()
+
+        marker = "Bitte Rubrik neu wählen."
+
+        assert marker not in editor_1.get('/notice/notice/edit')
+
+        change_category(gazette_app, '11', active=False)
+        assert marker in editor_1.get('/notice/notice/edit')
+
+
+def test_view_notice_edit_invalid_organization(gazette_app):
+    editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
+
+    with freeze_time("2017-11-01 11:00"):
+        manage = editor_1.get('/notices/drafted/new-notice')
+        manage.form['title'] = "Notice"
+        manage.form['organization'] = '200'
+        manage.form['category'] = '11'
+        manage.form['issues'] = ['2017-44', '2017-45']
+        manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
+        manage.form.submit()
+
+        marker = "Bitte Organisation neu wählen."
+
+        assert marker not in editor_1.get('/notice/notice/edit')
+
+        change_organization(gazette_app, '200', active=False)
+        assert marker in editor_1.get('/notice/notice/edit')
+
+
+def test_view_notice_edit_unrestricted(gazette_app):
+    editor_1 = Client(gazette_app)
+    login_editor_1(editor_1)
+
+    publisher = Client(gazette_app)
+    login_publisher(publisher)
+
+    admin = Client(gazette_app)
+    login_admin(admin)
+
+    then = "2017-11-01 11:00"
+    future = "2020-11-01 11:00"
+
+    with freeze_time(then):
+        manage = editor_1.get('/notices/drafted/new-notice')
+        manage.form['title'] = "Notice"
+        manage.form['organization'] = '200'
+        manage.form['category'] = '11'
+        manage.form['issues'] = ['2017-44', '2017-45']
+        manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
+        manage.form.submit()
+
+    # drafted
+    with freeze_time(future):
+        edit_notice_unrestricted(editor_1, 'notice', forbidden=True)
+        edit_notice_unrestricted(publisher, 'notice', forbidden=True)
+        edit_notice_unrestricted(admin, 'notice', title='unres_drafted')
+        assert 'unres_drafted' in editor_1.get('/notice/notice')
+
+        manage = admin.get('/notice/notice/edit_unrestricted')
+        assert "(Complaints)" in manage
+        assert "(Sikh Community)" in manage
+
+    # submitted
+    with freeze_time(then):
+        submit_notice(editor_1, 'notice')
+    with freeze_time(future):
+        edit_notice_unrestricted(editor_1, 'notice', forbidden=True)
+        edit_notice_unrestricted(publisher, 'notice', forbidden=True)
+        edit_notice_unrestricted(admin, 'notice', title='unres_submitted')
+        assert 'unres_submitted' in editor_1.get('/notice/notice')
+
+        manage = admin.get('/notice/notice/edit_unrestricted')
+        assert "(Complaints)" in manage
+        assert "(Sikh Community)" in manage
+
+    # rejected
+    with freeze_time(then):
+        reject_notice(publisher, 'notice')
+    with freeze_time(future):
+        edit_notice_unrestricted(editor_1, 'notice', forbidden=True)
+        edit_notice_unrestricted(publisher, 'notice', forbidden=True)
+        edit_notice_unrestricted(admin, 'notice', title='unres_rejected')
+        assert 'unres_rejected' in editor_1.get('/notice/notice')
+
+        manage = admin.get('/notice/notice/edit_unrestricted')
+        assert "(Complaints)" in manage
+        assert "(Sikh Community)" in manage
+
+    # accepted
+    with freeze_time(then):
+        submit_notice(editor_1, 'notice')
+        accept_notice(publisher, 'notice')
+    with freeze_time(future):
+        edit_notice_unrestricted(editor_1, 'notice', forbidden=True)
+        edit_notice_unrestricted(publisher, 'notice', forbidden=True)
+        edit_notice_unrestricted(admin, 'notice', title='unres_accepted')
+        assert 'unres_accepted' in editor_1.get('/notice/notice')
+
+        manage = admin.get('/notice/notice/edit_unrestricted')
+        assert "(Complaints)" in manage
+        assert "(Sikh Community)" in manage
+        assert "Diese Meldung wurde bereits angenommen!" in manage
+
+
 def test_view_notice_copy(gazette_app):
     admin, editor_1, editor_2, editor_3, publisher = login_users(gazette_app)
 
@@ -430,6 +893,9 @@ def test_view_notice_copy(gazette_app):
         manage.form['category'] = '11'
         manage.form['issues'] = ['2017-40']
         manage.form['text'] = "1. Oktober 2017"
+        manage.form['author_place'] = 'Govikon'
+        manage.form['author_name'] = 'State Chancellerist'
+        manage.form['author_date'] = '2019-01-01'
         manage.form.submit()
 
         submit_notice(editor_1, 'erneuerungswahlen')